Refactor key generation.
