+++ lib/index.js
@@ -3,13 +3,15 @@
  */
 'use strict';
 
+const _asymmetricKey = require('./asymmetric-key');
+const _constants = require('./constants');
+const _symmetricKey = require('./symmetric-key');
 const assert = require('assert-plus');
 const base64url = require('base64url-universal');
 const bedrock = require('bedrock');
 const brPackageManager = require('bedrock-package-manager');
 const crypto = require('crypto');
 const database = require('bedrock-mongodb');
-const {LDKeyPair} = require('crypto-ld');
 const {promisify} = require('util');
 const {util: {BedrockError}} = bedrock;
 
@@ -41,105 +43,28 @@ bedrock.events.on('bedrock-mongodb.ready', async () => {
  */
 
 /**
- * Generates a new symmetric key.
+ * Generates a new key.
  *
  * @param {Object} options - The options to use.
  * @param {string} options.keyId - The key ID to use.
  * @param {Object} options.operation - The KMS operation.
  *
- * @returns {Promise<Object>} An object containing `{id}`.
+ * @returns {Promise<Object>} Key information.
  */
 exports.generateKey = async ({keyId, operation}) => {
   assert.string(keyId, 'options.keyId');
   assert.object(operation, 'options.operation');
 
-  let key;
-
   const {invocationTarget: {type}} = operation;
-  if(type === 'AesKeyWrappingKey2019') {
-    // TODO: support other lengths?
-    key = {
-      id: keyId,
-      type,
-      secret: base64url.encode(crypto.randomBytes(32))
-    };
-  } else if(type === 'Sha256HmacKey2019') {
-    // TODO: support other hashes?
-    key = {
-      id: keyId,
-      type,
-      secret: base64url.encode(crypto.randomBytes(32))
-    };
-  } else {
-    throw new Error(`Unknown key type "${type}".`);
+  assert.string(type, 'options.operation.type');
+
+  if(_constants.keyTypes.asymmetric.has(type)) {
+    return _asymmetricKey.generateKeyPair({keyId, type});
+  } else if(_constants.keyTypes.symmetric.has(type)) {
+    return _symmetricKey.generateKey({keyId, type});
   }
 
-  // insert the key and get the updated record
-  const now = Date.now();
-  const meta = {created: now};
-  const record = {
-    id: database.hash(keyId),
-    meta,
-    key
-  };
-  try {
-    await database.collections.ssm.insert(record, database.writeOptions);
-    return {id: keyId};
-  } catch(e) {
-    if(!database.isDuplicateError(e)) {
-      throw e;
-    }
-    throw new BedrockError(
-      'Duplicate key identifier.',
-      'DuplicateError', {
-        public: true,
-        httpStatusCode: 409
-      }, e);
-  }
-};
-
-/**
- * Generates a new asymmetric key pair.
- *
- * @param {Object} options - The options to use.
- * @param {string} options.keyId - The key ID to use.
- * @param {Object} options.operation - The KMS operation.
- *
- * @returns {Promise<Object>} An object containing public key material.
- */
-exports.generateKeyPair = async ({keyId, operation}) => {
-  assert.string(keyId, 'options.keyId');
-  assert.object(operation, 'options.operation');
-
-  const now = Date.now();
-  const meta = {created: now};
-  const {invocationTarget: {type}} = operation;
-  const key = await LDKeyPair.generate({id: keyId, type});
-  // remove unused properties
-  delete key.controller;
-  delete key.owner;
-  delete key.passphrase;
-  const record = {
-    id: database.hash(keyId),
-    meta,
-    key
-  };
-
-  try {
-    await database.collections.ssm.insert(record, database.writeOptions);
-  } catch(e) {
-    if(!database.isDuplicateError(e)) {
-      throw e;
-    }
-    throw new BedrockError(
-      'Duplicate key identifier.',
-      'DuplicateError', {
-        public: true,
-        httpStatusCode: 409
-      }, e);
-  }
-
-  return key.publicNode();
+  throw new Error(`Unknown key type "${type}".`);
 };
 
 /**
