+++ test/mocha/01-symmetric-key.js
