+++ test/mocha/10-asymmetric-key.js
@@ -6,14 +6,14 @@
 const brSSM = require('bedrock-ssm-mongodb');
 const {util: {uuid}} = require('bedrock');
 
-describe('generateKeyPair API', () => {
-  it('returns a key id', async () => {
+describe('asymmetric keys', () => {
+  it('successfully generates a Ed25519VerificationKey2018', async () => {
     // uuid will be generated at the bedrock-kms-http layer
     const keyId = `https://example.com/ksm/ssm-v1/${uuid()}`;
     const controller = 'https://example.com/i/foo';
     const type = 'Ed25519VerificationKey2018';
     const invocationTarget = {id: keyId, type, controller};
-    const result = await brSSM.generateKeyPair(
+    const result = await brSSM.generateKey(
       {keyId, operation: {invocationTarget}});
     should.exist(result);
     result.should.be.an('object');
